+++ CutTheRope/GameMain/GameScene.Systems.cs
@@ -324,15 +324,15 @@ namespace CutTheRope.GameMain
             Timeline timeline = new Timeline().InitWithMaxKeyFramesOnTrack(3);
             if (gravityButton != null && !gravityNormal)
             {
-                timeline.AddKeyFrame(KeyFrame.MakePos(g.spider.x, g.spider.y, KeyFrame.TransitionType.FRAME_TRANSITION_EASE_OUT, 0));
-                timeline.AddKeyFrame(KeyFrame.MakePos(g.spider.x, g.spider.y + 50, KeyFrame.TransitionType.FRAME_TRANSITION_EASE_OUT, 0.3));
-                timeline.AddKeyFrame(KeyFrame.MakePos(g.spider.x, (double)(g.spider.y - SCREEN_HEIGHT), KeyFrame.TransitionType.FRAME_TRANSITION_EASE_IN, 1));
+                timeline.AddKeyFrame(KeyFrame.MakePos((int)g.spider.x, (int)g.spider.y, KeyFrame.TransitionType.FRAME_TRANSITION_EASE_OUT, 0));
+                timeline.AddKeyFrame(KeyFrame.MakePos((int)g.spider.x, (int)(g.spider.y + 50), KeyFrame.TransitionType.FRAME_TRANSITION_EASE_OUT, 0.3f));
+                timeline.AddKeyFrame(KeyFrame.MakePos((int)g.spider.x, (int)(double)(g.spider.y - SCREEN_HEIGHT), KeyFrame.TransitionType.FRAME_TRANSITION_EASE_IN, 1));
             }
             else
             {
-                timeline.AddKeyFrame(KeyFrame.MakePos(g.spider.x, g.spider.y, KeyFrame.TransitionType.FRAME_TRANSITION_EASE_OUT, 0));
-                timeline.AddKeyFrame(KeyFrame.MakePos(g.spider.x, g.spider.y - 50, KeyFrame.TransitionType.FRAME_TRANSITION_EASE_OUT, 0.3));
-                timeline.AddKeyFrame(KeyFrame.MakePos(g.spider.x, (double)(g.spider.y + SCREEN_HEIGHT), KeyFrame.TransitionType.FRAME_TRANSITION_EASE_IN, 1));
+                timeline.AddKeyFrame(KeyFrame.MakePos((int)g.spider.x, (int)g.spider.y, KeyFrame.TransitionType.FRAME_TRANSITION_EASE_OUT, 0));
+                timeline.AddKeyFrame(KeyFrame.MakePos((int)g.spider.x, (int)(g.spider.y - 50), KeyFrame.TransitionType.FRAME_TRANSITION_EASE_OUT, 0.3f));
+                timeline.AddKeyFrame(KeyFrame.MakePos((int)g.spider.x, (int)(double)(g.spider.y + SCREEN_HEIGHT), KeyFrame.TransitionType.FRAME_TRANSITION_EASE_IN, 1));
             }
             timeline.AddKeyFrame(KeyFrame.MakeRotation(0, KeyFrame.TransitionType.FRAME_TRANSITION_LINEAR, 0));
             timeline.AddKeyFrame(KeyFrame.MakeRotation(RND_RANGE(-120, 120), KeyFrame.TransitionType.FRAME_TRANSITION_LINEAR, 1));
@@ -401,15 +401,15 @@ namespace CutTheRope.GameMain
             Timeline timeline = new Timeline().InitWithMaxKeyFramesOnTrack(3);
             if (gravityButton != null && !gravityNormal)
             {
-                timeline.AddKeyFrame(KeyFrame.MakePos(sg.spider.x, sg.spider.y - 10, KeyFrame.TransitionType.FRAME_TRANSITION_EASE_OUT, 0));
-                timeline.AddKeyFrame(KeyFrame.MakePos(sg.spider.x, sg.spider.y + 70, KeyFrame.TransitionType.FRAME_TRANSITION_EASE_OUT, 0.3));
-                timeline.AddKeyFrame(KeyFrame.MakePos(sg.spider.x, (double)(sg.spider.y - SCREEN_HEIGHT), KeyFrame.TransitionType.FRAME_TRANSITION_EASE_IN, 1));
+                timeline.AddKeyFrame(KeyFrame.MakePos((int)sg.spider.x, (int)(sg.spider.y - 10), KeyFrame.TransitionType.FRAME_TRANSITION_EASE_OUT, 0));
+                timeline.AddKeyFrame(KeyFrame.MakePos((int)sg.spider.x, (int)(sg.spider.y + 70), KeyFrame.TransitionType.FRAME_TRANSITION_EASE_OUT, 0.3f));
+                timeline.AddKeyFrame(KeyFrame.MakePos((int)sg.spider.x, (int)(double)(sg.spider.y - SCREEN_HEIGHT), KeyFrame.TransitionType.FRAME_TRANSITION_EASE_IN, 1));
             }
             else
             {
-                timeline.AddKeyFrame(KeyFrame.MakePos(sg.spider.x, sg.spider.y - 10, KeyFrame.TransitionType.FRAME_TRANSITION_EASE_OUT, 0));
-                timeline.AddKeyFrame(KeyFrame.MakePos(sg.spider.x, sg.spider.y - 70, KeyFrame.TransitionType.FRAME_TRANSITION_EASE_OUT, 0.3));
-                timeline.AddKeyFrame(KeyFrame.MakePos(sg.spider.x, (double)(sg.spider.y + SCREEN_HEIGHT), KeyFrame.TransitionType.FRAME_TRANSITION_EASE_IN, 1));
+                timeline.AddKeyFrame(KeyFrame.MakePos((int)sg.spider.x, (int)(sg.spider.y - 10), KeyFrame.TransitionType.FRAME_TRANSITION_EASE_OUT, 0));
+                timeline.AddKeyFrame(KeyFrame.MakePos((int)sg.spider.x, (int)(sg.spider.y - 70), KeyFrame.TransitionType.FRAME_TRANSITION_EASE_OUT, 0.3f));
+                timeline.AddKeyFrame(KeyFrame.MakePos((int)sg.spider.x, (int)(double)(sg.spider.y + SCREEN_HEIGHT), KeyFrame.TransitionType.FRAME_TRANSITION_EASE_IN, 1));
             }
             image.AddTimelinewithID(timeline, 0);
             image.PlayTimeline(0);
